cambios finales mios

+++ PHP/modificararticulo.php
@@ -16,23 +16,29 @@ $resultado = $conn->query($sql);
 while($resul = mysqli_fetch_array($resultado)){
     $img = $resul[0];
 }
-$ruta_carpeta="../assets/img/";
-$imagenborrar = $ruta_carpeta.$img;
-$nombre_imagen = $_POST['cod'].".".pathinfo($_FILES['img']['name'],PATHINFO_EXTENSION);
-$ruta_guardar_archivo=$ruta_carpeta.$nombre_imagen;
-if(unlink($imagenborrar)){
-    if(move_uploaded_file($_FILES['img']['tmp_name'],$ruta_guardar_archivo))
-{
-    $articulo->setImagen(strip_tags($nombre_imagen));
-}
-else
-{
-    echo "no se pudo";
-}
-$articulo->ActualizarArticulo($articulo);
-echo "Articulo modificado Exitosamente";
-}
-else{
-    echo "no se puedo compadre";
+if(isset($_POST['img'])){
+    $articulo->setImagen(strip_tags($img));
+    $articulo->ActualizarArticulo($articulo);
+
+}else{
+    $ruta_carpeta="../assets/img/";
+    $imagenborrar = $ruta_carpeta.$img;
+    $nombre_imagen = $_POST['cod'].".".pathinfo($_FILES['img']['name'],PATHINFO_EXTENSION);
+    $ruta_guardar_archivo=$ruta_carpeta.$nombre_imagen;
+    if(unlink($imagenborrar)){
+        if(move_uploaded_file($_FILES['img']['tmp_name'],$ruta_guardar_archivo))
+    {
+        $articulo->setImagen(strip_tags($nombre_imagen));
+    }
+    else
+    {
+        echo "no se pudo";
+    }
+    $articulo->ActualizarArticulo($articulo);
+    echo "Articulo modificado Exitosamente";
+    }
+    else{
+        echo "no se pudo";
+    }
 }
 ?>
